move Channel
